Extract config from playwright
