+++ packages/tests/playwright.config.ts
@@ -1,19 +1,8 @@
 import { PlaywrightTestConfig, defineConfig, devices } from "@playwright/test";
-import dotenv from "dotenv";
 import path from "path";
-
-dotenv.config();
+import { config as envConfig } from "./config";
 
 function createPlaywrightConfig(): PlaywrightTestConfig {
-  const envConfig = {
-    app: process.env.APP_NAME || "",
-    CI: process.env.CI === "true",
-  };
-
-  if (!envConfig.app) {
-    throw new Error("APP_NAME is required in .env file");
-  }
-
   return {
     timeout: 16 * 1000,
     expect: {
